refactor(stage3): consolidate BacklogPaper parser into woeai.wechat.backlog
Candidate B, slice 1. BacklogPaper + parse_backlog_papers were duplicated
across markdown_to_rtd.py (6-field) and wechat_draft.py (7-field); both now
import the 7-field superset from woeai/wechat/backlog.py.

- woeai/wechat/backlog.py: BacklogPaper (superset), parse_backlog_papers,
  read_backlog_item, read_backlog_publication_refs (single implementation).
- markdown_to_rtd.py: drops its 6-field copy, imports the shared one.
- wechat_draft.py: drops its 7-field copy + 3 backlog readers, imports them.
  Both scripts re-export so test_wechat_tools (draft.BacklogPaper(...)) is
  unchanged.
- Both entry scripts now self-bootstrap sys.path for 'import woeai'.

55 tests green.
